remove forward_next and put back into forward
