+++ mujoco_warp/_src/forward.py
@@ -22,16 +22,13 @@ from mujoco_warp._src import constraint
 from mujoco_warp._src import derivative
 from mujoco_warp._src import history
 from mujoco_warp._src import island
+from mujoco_warp._src import math
 from mujoco_warp._src import passive
 from mujoco_warp._src import sensor
 from mujoco_warp._src import sleep
 from mujoco_warp._src import smooth
 from mujoco_warp._src import solver
 from mujoco_warp._src import util_misc
-from mujoco_warp._src.forward_next import next_activation
-from mujoco_warp._src.forward_next import next_position
-from mujoco_warp._src.forward_next import next_time
-from mujoco_warp._src.forward_next import next_velocity
 from mujoco_warp._src.support import next_act
 from mujoco_warp._src.support import xfrc_accumulate
 from mujoco_warp._src.types import MJ_MINVAL
@@ -50,6 +47,76 @@ from mujoco_warp._src.warp_util import cache_kernel
 from mujoco_warp._src.warp_util import event_scope
 
 wp.set_module_options({"enable_backward": False})
+
+
+@wp.func
+def next_position(
+  # Data in:
+  qpos_in: wp.array2d[float],
+  # In:
+  jnttype: int,
+  qpos_adr: int,
+  timestep: float,
+  worldid: int,
+  qvel_lin: wp.vec3,
+  qvel_ang: wp.vec3,
+  # Data out:
+  qpos_out: wp.array2d[float],
+):
+  """Per-joint semi-implicit position integration from velocity values."""
+  if jnttype == JointType.FREE:
+    qpos_pos = wp.vec3(qpos_in[worldid, qpos_adr], qpos_in[worldid, qpos_adr + 1], qpos_in[worldid, qpos_adr + 2])
+    qpos_new = qpos_pos + timestep * qvel_lin
+
+    qpos_quat = wp.quat(
+      qpos_in[worldid, qpos_adr + 3],
+      qpos_in[worldid, qpos_adr + 4],
+      qpos_in[worldid, qpos_adr + 5],
+      qpos_in[worldid, qpos_adr + 6],
+    )
+    qpos_quat_new = math.quat_integrate(qpos_quat, qvel_ang, timestep)
+
+    qpos_out[worldid, qpos_adr + 0] = qpos_new[0]
+    qpos_out[worldid, qpos_adr + 1] = qpos_new[1]
+    qpos_out[worldid, qpos_adr + 2] = qpos_new[2]
+    qpos_out[worldid, qpos_adr + 3] = qpos_quat_new[0]
+    qpos_out[worldid, qpos_adr + 4] = qpos_quat_new[1]
+    qpos_out[worldid, qpos_adr + 5] = qpos_quat_new[2]
+    qpos_out[worldid, qpos_adr + 6] = qpos_quat_new[3]
+
+  elif jnttype == JointType.BALL:
+    qpos_quat = wp.quat(
+      qpos_in[worldid, qpos_adr + 0],
+      qpos_in[worldid, qpos_adr + 1],
+      qpos_in[worldid, qpos_adr + 2],
+      qpos_in[worldid, qpos_adr + 3],
+    )
+    qpos_quat_new = math.quat_integrate(qpos_quat, qvel_ang, timestep)
+
+    qpos_out[worldid, qpos_adr + 0] = qpos_quat_new[0]
+    qpos_out[worldid, qpos_adr + 1] = qpos_quat_new[1]
+    qpos_out[worldid, qpos_adr + 2] = qpos_quat_new[2]
+    qpos_out[worldid, qpos_adr + 3] = qpos_quat_new[3]
+
+  else:  # HINGE / SLIDE
+    qpos_out[worldid, qpos_adr] = qpos_in[worldid, qpos_adr] + timestep * qvel_lin[0]
+
+
+@wp.func
+def next_velocity(
+  # Model:
+  opt_timestep: wp.array[float],
+  # Data in:
+  qvel_in: wp.array2d[float],
+  qacc_in: wp.array2d[float],
+  # In:
+  worldid: int,
+  dofid: int,
+  qacc_scale_in: float,
+) -> float:
+  # returns the updated velocity (not written) so the backward integrator can hold it as a local
+  timestep = opt_timestep[worldid % opt_timestep.shape[0]]
+  return qvel_in[worldid, dofid] + qacc_scale_in * qacc_in[worldid, dofid] * timestep
 
 
 @wp.kernel
@@ -86,7 +153,7 @@ def _next_position(
   else:  # HINGE / SLIDE
     qvel_lin = wp.vec3(qvel[dof_adr] * qvel_scale_in, 0.0, 0.0)
 
-  next_position(jnttype, qpos_adr, timestep, qpos_in, worldid, qvel_lin, qvel_ang, qpos_out)
+  next_position(qpos_in, jnttype, qpos_adr, timestep, worldid, qvel_lin, qvel_ang, qpos_out)
 
 
 @wp.kernel
@@ -102,7 +169,7 @@ def _next_velocity(
   qvel_out: wp.array2d[float],
 ):
   worldid, dofid = wp.tid()
-  qvel_out[worldid, dofid] = next_velocity(worldid, dofid, opt_timestep, qvel_in, qacc_in, qacc_scale_in)
+  qvel_out[worldid, dofid] = next_velocity(opt_timestep, qvel_in, qacc_in, worldid, dofid, qacc_scale_in)
 
 
 @wp.kernel
@@ -128,25 +195,68 @@ def _next_activation(
   act_out: wp.array2d[float],
 ):
   worldid, uid = wp.tid()
-  next_activation(
-    worldid,
-    uid,
-    opt_timestep,
-    actuator_dyntype,
-    actuator_actadr,
-    actuator_actnum,
-    actuator_actlimited,
-    actuator_dynprm,
-    actuator_gainprm,
-    actuator_biasprm,
-    actuator_actrange,
-    act_in,
-    act_dot_in,
-    actuator_velocity_in,
-    act_dot_scale,
-    limit,
-    act_out,
-  )
+  opt_timestep_id = worldid % opt_timestep.shape[0]
+  actuator_dynprm_id = worldid % actuator_dynprm.shape[0]
+  actuator_actrange_id = worldid % actuator_actrange.shape[0]
+  actuator_gainprm_id = worldid % actuator_gainprm.shape[0]
+  actuator_biasprm_id = worldid % actuator_biasprm.shape[0]
+
+  actadr = actuator_actadr[uid]
+  actnum = actuator_actnum[uid]
+  dyntype = actuator_dyntype[uid]
+
+  if dyntype == DynType.DCMOTOR:
+    dynprm = actuator_dynprm[actuator_dynprm_id, uid]
+    gainprm = actuator_gainprm[actuator_gainprm_id, uid]
+    biasprm = actuator_biasprm[actuator_biasprm_id, uid]
+    slots = util_misc.dcmotor_slots(dynprm, gainprm)
+
+    for j in range(actadr, actadr + actnum):
+      offset = j - actadr
+      act = act_in[worldid, j]
+      act_dot = act_dot_in[worldid, j]
+
+      if offset == slots[4]:  # current
+        R = gainprm[0]
+        te = wp.max(MJ_MINVAL, dynprm[0])
+        act = act + act_dot * te * (1.0 - wp.exp(-opt_timestep[opt_timestep_id] / te))
+      elif offset == slots[3]:  # bristle
+        F_C = biasprm[3]
+        F_S = biasprm[4]
+        v_S = biasprm[5]
+        sigma0 = dynprm[5]
+        velocity = actuator_velocity_in[worldid, uid]
+        g = util_misc.lugre_stribeck(velocity, F_C, F_S, v_S)
+
+        a = -sigma0 * wp.abs(velocity) / wp.max(MJ_MINVAL, g)
+        h = opt_timestep[opt_timestep_id]
+        exp_ah = wp.exp(a * h)
+        int_h = h
+        if wp.abs(a) > MJ_MINVAL:
+          int_h = (exp_ah - 1.0) / a
+        act = exp_ah * act + int_h * velocity
+      elif offset == slots[1]:  # integral
+        act = act + act_dot * opt_timestep[opt_timestep_id]
+        Imax = dynprm[8]
+        if Imax > 0.0:
+          act = wp.clamp(act, -Imax, Imax)
+      else:  # temperature and slew
+        act = act + act_dot * opt_timestep[opt_timestep_id]
+
+      act_out[worldid, j] = act
+  else:
+    for j in range(actadr, actadr + actnum):
+      act = next_act(
+        opt_timestep[opt_timestep_id],
+        dyntype,
+        actuator_dynprm[actuator_dynprm_id, uid],
+        actuator_actrange[actuator_actrange_id, uid],
+        act_in[worldid, j],
+        act_dot_in[worldid, j],
+        act_dot_scale,
+        limit and actuator_actlimited[uid],
+      )
+      act_out[worldid, j] = act
 
 
 @wp.kernel
@@ -169,7 +279,7 @@ def _next_time(
   time_out: wp.array[float],
 ):
   worldid = wp.tid()
-  next_time(worldid, opt_timestep, time_in, time_out)
+  time_out[worldid] = time_in[worldid] + opt_timestep[worldid % opt_timestep.shape[0]]
 
   nefc = nefc_in[worldid]
   if nefc > njmax_in:
